+++ Configurations/Configuration.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Dalamud.Configuration;
+using FFXIVClientStructs;
 using Lumina.Excel.Sheets;
 using Newtonsoft.Json;
 using static ActionTimelineReplacement.Configurations.Configuration;
@@ -13,14 +14,14 @@ namespace ActionTimelineReplacement.Configurations;
 [Serializable]
 public class Configuration : IPluginConfiguration
 {
-    public class ReplacementSet(string name, bool enabled, int priority, Dictionary<uint, ActionReplacementConfig> actionreplacements, Dictionary<uint, ActionCastVFXReplacementConfig> actioncastvfxreplacements, Dictionary<uint, MountReplacementConfig> mountreplacements, Dictionary<uint, TiltReplacementConfig> tiltreplacements)
+    public class ReplacementSet (string name, bool enabled, int priority, Dictionary<uint, ActionReplacementConfig> actionreplacements, Dictionary<uint, ActionCastVFXReplacementConfig> actioncastvfxreplacements, Dictionary<uint, MountReplacementConfig> mountreplacements, Dictionary<uint, TiltReplacementConfig> tiltreplacements)
     {
         public string Name = name;
 
         public bool Enabled = enabled;
 
         public int Priority = priority;
-
+        
         public Dictionary<uint, ActionReplacementConfig> ActionReplacements { get; set; } = actionreplacements;
         public Dictionary<uint, ActionCastVFXReplacementConfig> ActionCastVFXReplacements { get; set; } = actioncastvfxreplacements;
         public Dictionary<uint, MountReplacementConfig> MountReplacements { get; set; } = mountreplacements;
@@ -74,5 +75,4 @@ public class Configuration : IPluginConfiguration
     {
         Service.PluginInterface.SavePluginConfig(this);
     }
-
 }
