+++ Configurations/Interfaces/IReplacementsManager.cs
@@ -1,0 +1,8 @@
+﻿namespace ActionTimelineReplacement.Configurations.Interfaces
+{
+    internal interface IReplacementsManager
+    {
+        void thisManager();
+
+    }
+}
