+++ Hookers/HookersData/ActionCastTimelineData.cs
@@ -1,0 +1,9 @@
+﻿using System.Runtime.InteropServices;
+
+namespace ActionTimelineReplacement.Hookers;
+
+[StructLayout(LayoutKind.Explicit)]
+public struct ActionCastTimelineData
+{
+    //FILLMEOUT
+}
