+++ Hookers/HookersData/StatusHitEffectData.cs
