+++ Windows/ConfigWindow.cs
@@ -5,6 +5,7 @@ using System.Numerics;
 using System.Runtime.InteropServices;
 using ActionTimelineReplacement.Configurations;
 using ActionTimelineReplacement.Hookers;
+using ActionTimelineReplacement.Windows.SubSheets;
 using Dalamud.Interface;
 using Dalamud.Interface.Colors;
 using Dalamud.Interface.GameFonts;
@@ -18,10 +19,6 @@ namespace ActionTimelineReplacement.Windows;
 
 public sealed class ConfigWindow : Window
 {
-    private float _ActionitemWidth;
-    private float _MountitemWidth;
-    private float _TiltitemWidth;
-    private float _BGMitemWidth;
 
     private readonly FileDialogManager _dialogManager = new()
     {
@@ -132,7 +129,7 @@ public sealed class ConfigWindow : Window
                         _dialogManager.SaveFileDialog("Save", ".json", Service.Config.ReplacementSets[i].Name, ".json", (b, file) =>
                         {
                             if (!b) return;
-                            //FIX
+                            //FIXME
                             //set.Save(file);
                         });
                     }
@@ -228,632 +225,79 @@ public sealed class ConfigWindow : Window
             Service.Config.Save();
         }
     }
-    
+
+    private Dictionary<string, List<string>> _AllHeaders = [];
+    private Dictionary<string, float> _AllItemWidths = [];
+
+    //TOSETUP: Add new search here
+    private string _searchAction = string.Empty;
+    private string _searchMount = string.Empty;
+    private string _searchTiltParam = string.Empty;
+
 
     private void DrawBodySheets()
     {
+        if (!_AllItemWidths.ContainsKey("Action"))
+        {
+            //TOSETUP: Add new headers here
+            _AllHeaders.Add("Action", ["Cast Vfx", "Start timeline", "End timeline", "Hit timeline"]);
+            //_AllHeaders.Add("Mount", ["RideBGM", "TiltGround", "TiltFlySwim", "TiltParam3", "TiltParam4", "FlyUpDownTilt", "Unk2", "Unk3", "Unk4", "MountCustomize", "Unk5", "SwimAnimSpeed"]);
+            //_AllHeaders.Add("TiltParam", ["TiltRate", "RotOriginOffset", "MaxAngle", "Unknown3", "Unknown4", "RotReverse"]);
+            foreach (var headerkey in _AllHeaders.Keys)
+            {
+                _AllItemWidths.Add(headerkey, 0f);
+            }
 
+        }
         using var child = ImRaii.Child("BodySheets", new Vector2(-1f, -1f), false);
         if (!child) return;
         if (_activeSet is null) return;
 
         ImGui.SetCursorPosY(ImGui.GetCursorPosY() + 10 * Scale);
-
-        
-        if (ImGui.CollapsingHeader($"Action"))
+        foreach (var mainkey in _AllHeaders.Keys)
         {
-            using (ImRaii.PushFont(GetFont(18)))
+            if (ImGui.CollapsingHeader(mainkey))
             {
-
-                if (Service.Config.AdvancedMode)
+                using (ImRaii.PushFont(GetFont(18)))
                 {
-                    if (_ActionitemWidth == 0)
+                    if (Service.Config.AdvancedMode)
                     {
-                        ImGui.SameLine();
-                        ImGui.Text(" Cast Vfx Start timeline End timeline Hit timeline");
-                    }
-                    else
-                    {
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _ActionitemWidth * 4 / 4);
-                        ImGui.Text("Cast Vfx");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _ActionitemWidth * 3 / 4);
-                        ImGui.Text("Start timeline");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _ActionitemWidth * 2 / 4);
-                        ImGui.Text("End timeline");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _ActionitemWidth * 1 / 4);
-                        ImGui.Text("Hit timeline");
-                    }
-                }
-            }
-
-            using (var subList = ImRaii.Child("ActionSubList", new Vector2(-1f, (float)(_activeSet.ActionReplacements.Count()*30)+30), false))
-            {
-                if (subList)
-                {
-
-                    foreach (var key in _activeSet.ActionReplacements.Keys)
-                    {
-                        var replacement = _activeSet.ActionReplacements[key].Replacement;
-
-                        if (ImGui.Checkbox("##" + key, ref _activeSet.ActionReplacements[key].Enabled))
+                        if (_AllItemWidths[mainkey] == 0)
                         {
-                            Methods.SetupAction(key);
-                            Service.Config.Save();
-                        }
-                        ImGui.SameLine();
-                        if (ImGui.Button(" - ##" + key))
-                        {
-                            _activeSet.ActionReplacements.Remove(key);
-                        }
-
-                        ImGui.SameLine();
-                        ImGui.Text($"#{key:D5}");
-
-                        ImGui.SameLine();
-                        if (_ActionitemWidth != 0 && Service.Config.AdvancedMode)
-                        {
-                            var widthRest = ImGui.GetWindowWidth() - _ActionitemWidth - ImGui.GetCursorPosX() - 5 * Scale;
-                            ImGui.PushTextWrapPos(Math.Max(widthRest, 60 * Scale) + ImGui.GetCursorPosX());
-                        }
-                        ImGui.TextWrapped(ActionReplacementsManager.GetName(key));
-                        if (_ActionitemWidth != 0 && Service.Config.AdvancedMode)
-                        {
-                            ImGui.PopTextWrapPos();
-                        }
-
-                        if (!Service.Config.AdvancedMode)
-                        {
-                            continue;
-                        }
-
-                        ImGui.SameLine();
-                        if (_ActionitemWidth != 0)
-                        {
-                            ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _ActionitemWidth);
-                        }
-
-                        var startwidth = ImGui.GetCursorPosX();
-                        DrawItem("Cast", ref replacement.CastVfx, i => i.CastVfx);
-                        ImGui.SameLine();
-
-                        DrawItem("Start", ref replacement.AnimationStart, i => i.AnimationStart);
-
-                        ImGui.SameLine();
-                        DrawItem("End", ref replacement.AnimationEnd, i => i.AnimationEnd);
-
-                        ImGui.SameLine();
-                        DrawItem("Hit", ref replacement.ActionTimelineHit, i => i.ActionTimelineHit);
-
-                        ImGui.SameLine();
-                        _ActionitemWidth = ImGui.GetCursorPosX() - startwidth;
-                        ImGui.NewLine();
-                        continue;
-
-                        void DrawItem(string name, ref ushort value,
-                            Func<Configurations.ActionReplacement, ushort> getDefault)
-                        {
-                            ImGui.SetNextItemWidth(60 * Scale);
-                            int relay = value;
-                            if (ImGui.DragInt("##" + name + key, ref relay))
-                            {
-                                value = (ushort)relay;
-                                Methods.SetupAction(key);
-                                Service.Config.Save();
-                            }
-
                             ImGui.SameLine();
-                            using (ImRaii.PushFont(UiBuilder.IconFont))
+                            var headerstring = " ";
+                            for (int i = 0; i < _AllHeaders[mainkey].Count; i++)
                             {
-                                if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
-                                {
-                                    value = getDefault(ActionReplacementsManager.GetOriginalReplacement(key));
-                                    Methods.SetupAction(key);
-                                    Service.Config.Save();
-                                }
+                                if (i > 0) headerstring += " ";
+                                headerstring += _AllHeaders[mainkey][i];
                             }
+                            ImGui.Text(headerstring);
                         }
-                    }
-
-                    const string searchActionsPopup = "Search actions";
-                    if (ImGui.Button(" + "))
-                    {
-                        ImGui.OpenPopup(searchActionsPopup);
-                    }
-
-                    using var searchPopup = ImRaii.Popup(searchActionsPopup);
-                    if (searchPopup)
-                    {
-                        var width = 200 * Scale;
-
-                        ImGui.SetNextItemWidth(width);
-                        ImGui.InputText("##Search Action", ref _searchAction, 256);
-
-                        using var popUpChild = ImRaii.Child(searchActionsPopup, new Vector2(width, 200 * Scale), true);
-                        foreach (var pair in ActionReplacementsManager.ActionNames.OrderBy(i =>
-                                {
-                                    if (string.IsNullOrEmpty(_searchAction)) return 0;
-                                    return Math.Min(ScoreString(i.Value, _searchAction),
-                                        ScoreString(i.Key.ToString(), _searchAction));
-                                }))
+                        else
                         {
-                            if (ImGui.Selectable($"#{pair.Key:D5} {pair.Value}"))
+                            for (int i = 0; i < _AllHeaders[mainkey].Count; i++)
                             {
-                                var original = ActionReplacementsManager.GetOriginalReplacement(pair.Key);
-                                _activeSet.ActionReplacements[pair.Key] =
-                                    new ActionReplacementConfig(new Configurations.ActionReplacement(
-                                            original.AnimationStart,
-                                            original.AnimationEnd,
-                                            original.ActionTimelineHit,
-                                            original.CastVfx),
-                                        false);
-                                Service.Config.Save();
+                                ImGui.SameLine();
+                                ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _AllItemWidths[mainkey] * (_AllHeaders[mainkey].Count - i) / _AllHeaders[mainkey].Count);
+                                ImGui.Text(_AllHeaders[mainkey][i]);
                             }
                         }
                     }
                 }
+                switch (mainkey)
+                {
+                    //TOSETUP: Add new case here to call the subsheet
+                    case "Action":
+                        ActionSubSheet.Draw(mainkey, ref _activeSet, ref _AllItemWidths, _searchAction);
+                        break;
+                }
+
             }
         }
 
-        if (ImGui.CollapsingHeader($"Mount"))
-        {
-            using (ImRaii.PushFont(GetFont(18)))
-            {
-
-                if (Service.Config.AdvancedMode)
-                {
-                    if (_MountitemWidth == 0)
-                    {
-                        ImGui.SameLine();
-                        ImGui.Text(" RideBGM TiltGround TiltFlySwim TiltParam3 TiltParam4 FlyUpDownTilt Unk2 Unk3 Unk4 MountCustomize Unk5 SwimAnimSpeed");
-                    }
-                    else
-                    {
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 12 / 12);
-                        ImGui.Text("RideBGM");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 11 / 12);
-                        ImGui.Text("TiltGround");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 10 / 12);
-                        ImGui.Text("TiltFlySwim");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 9 / 12);
-                        ImGui.Text("TiltParam3");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 8 / 12);
-                        ImGui.Text("TiltParam4");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 7 / 12);
-                        ImGui.Text("FlyUpDownTilt");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 6 / 12);
-                        ImGui.Text("Unk2");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 5 / 12);
-                        ImGui.Text("Unk3");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 4 / 12);
-                        ImGui.Text("Unk4");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 3 / 12);
-                        ImGui.Text("MountCustomize");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 2 / 12);
-                        ImGui.Text("Unk5");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth * 1 / 12);
-                        ImGui.Text("SwimAnimSpeed");
-                    }
-                }
-            }
-
-            using (var subList = ImRaii.Child("MountSubList", new Vector2(-1f, (float)(_activeSet.MountReplacements.Count() * 30) + 30), false))
-            {
-                if (subList)
-                {
-                    foreach (var key in _activeSet.MountReplacements.Keys)
-                    {
-                        var replacement = _activeSet.MountReplacements[key].Replacement;
-
-                        if (ImGui.Checkbox("##" + key, ref _activeSet.MountReplacements[key].Enabled))
-                        {
-                            Methods.SetupMount(key);
-                            Service.Config.Save();
-                        }
-
-                        ImGui.SameLine();
-                        if (ImGui.Button(" - ##" + key))
-                        {
-                            _activeSet.MountReplacements.Remove(key);
-                        }
-
-                        ImGui.SameLine();
-                        ImGui.Text($"#{key:D5}");
-
-                        ImGui.SameLine();
-                        if (_MountitemWidth != 0 && Service.Config.AdvancedMode)
-                        {
-                            var widthRest = ImGui.GetWindowWidth() - _MountitemWidth - ImGui.GetCursorPosX() - 5 * Scale;
-                            ImGui.PushTextWrapPos(Math.Max(widthRest, 60 * Scale) + ImGui.GetCursorPosX());
-                        }
-                        ImGui.TextWrapped(MountReplacementsManager.GetName(key));
-                        if (_MountitemWidth != 0 && Service.Config.AdvancedMode)
-                        {
-                            ImGui.PopTextWrapPos();
-                        }
-
-                        if (!Service.Config.AdvancedMode)
-                        {
-                            continue;
-                        }
-
-                        ImGui.SameLine();
-                        if (_MountitemWidth != 0)
-                        {
-                            ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _MountitemWidth);
-                        }
-
-                        var startwidth = ImGui.GetCursorPosX();
-                        DrawItem("RideBGM", ref replacement.RideBGM, i => i.RideBGM);
-                        ImGui.SameLine();
-
-                        DrawItem("TiltGround", ref replacement.TiltParam1, i => i.TiltParam1);
-
-                        ImGui.SameLine();
-                        DrawItem("TiltFlySwim", ref replacement.TiltParam2, i => i.TiltParam2);
-
-                        ImGui.SameLine();
-                        DrawItem("Tilt3", ref replacement.TiltParam3, i => i.TiltParam3);
-
-                        ImGui.SameLine();
-                        DrawItem("Tilt4", ref replacement.TiltParam4, i => i.TiltParam4);
-
-                        ImGui.SameLine();
-                        DrawItem("FlyUpDownTilt", ref replacement.Unk1, i => i.Unk1);
-
-                        ImGui.SameLine();
-                        DrawItem("Unk2", ref replacement.Unk2, i => i.Unk2);
-
-                        ImGui.SameLine();
-                        DrawItem("Unk3", ref replacement.Unk3, i => i.Unk3);
-
-                        ImGui.SameLine();
-                        DrawItem("Unk4", ref replacement.Unk4, i => i.Unk4);
-
-                        ImGui.SameLine();
-                        DrawItem("MountCustomize", ref replacement.MountCustomize, i => i.MountCustomize);
-
-                        ImGui.SameLine();
-                        DrawItem("Unk5", ref replacement.Unk5, i => i.Unk5);
-
-                        ImGui.SameLine();
-                        DrawItem("SwimAnimSpeed", ref replacement.Unk6, i => i.Unk6);
-
-                        ImGui.SameLine();
-                        _MountitemWidth = ImGui.GetCursorPosX() - startwidth;
-                        ImGui.NewLine();
-                        continue;
-
-                        void DrawItem(string name, ref ushort value,
-                            Func<Configurations.MountReplacement, ushort> getDefault)
-                        {
-                            ImGui.SetNextItemWidth(60 * Scale);
-                            int relay = value;
-                            if (ImGui.DragInt("##" + name + key, ref relay))
-                            {
-                                value = (ushort)relay;
-                                Methods.SetupMount(key);
-                                Service.Config.Save();
-                            }
-
-                            ImGui.SameLine();
-                            using (ImRaii.PushFont(UiBuilder.IconFont))
-                            {
-                                if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
-                                {
-                                    value = getDefault(MountReplacementsManager.GetOriginalReplacement(key));
-                                    Methods.SetupMount(key);
-                                    Service.Config.Save();
-                                }
-                            }
-                        }
-                    }
-
-                    const string searchMountsPopup = "Search mounts";
-                    if (ImGui.Button(" + "))
-                    {
-                        ImGui.OpenPopup(searchMountsPopup);
-                    }
-
-                    using var searchPopup = ImRaii.Popup(searchMountsPopup);
-                    if (searchPopup)
-                    {
-                        var width = 200 * Scale;
-
-                        ImGui.SetNextItemWidth(width);
-                        ImGui.InputText("##Search Action", ref _searchMount, 256);
-
-                        using var popUpChild = ImRaii.Child(searchMountsPopup, new Vector2(width, 200 * Scale), true);
-                        foreach (var pair in MountReplacementsManager.MountNames.OrderBy(i =>
-                        {
-                            if (string.IsNullOrEmpty(_searchMount)) return 0;
-                            return Math.Min(ScoreString(i.Value, _searchMount),
-                                ScoreString(i.Key.ToString(), _searchMount));
-                        }))
-                        {
-                            if (ImGui.Selectable($"#{pair.Key:D5} {pair.Value}"))
-                            {
-                                var original = MountReplacementsManager.GetOriginalReplacement(pair.Key);
-                                _activeSet.MountReplacements[pair.Key] =
-                                    new MountReplacementConfig(new Configurations.MountReplacement(
-                                            original.RideBGM,
-                                            original.TiltParam1,
-                                            original.TiltParam2,
-                                            original.TiltParam3,
-                                            original.TiltParam4,
-                                            original.Unk1,
-                                            original.Unk2,
-                                            original.Unk3,
-                                            original.Unk4,
-                                            original.MountCustomize,
-                                            original.Unk5,
-                                            original.Unk6),
-                                        false);
-                                Service.Config.Save();
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        if (ImGui.CollapsingHeader($"TiltParam"))
-        {
-            using (ImRaii.PushFont(GetFont(18)))
-            {
-
-                if (Service.Config.AdvancedMode)
-                {
-                    if (_TiltitemWidth == 0)
-                    {
-                        ImGui.SameLine();
-                        ImGui.Text("TiltRate RotOriginOffset MaxAngle Unknown3 Unknown4 RotReverse");
-                    }
-                    else
-                    {
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _TiltitemWidth * 6 / 6);
-                        ImGui.Text("TiltRate");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _TiltitemWidth * 5 / 6);
-                        ImGui.Text("RotOriginOffset");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _TiltitemWidth * 4 / 6);
-                        ImGui.Text("MaxAngle");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _TiltitemWidth * 3 / 6);
-                        ImGui.Text("Unknown3");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _TiltitemWidth * 2 / 6);
-                        ImGui.Text("Unknown4");
-
-                        ImGui.SameLine();
-                        ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _TiltitemWidth * 1 / 6);
-                        ImGui.Text("RotReverse");
-                    }
-                }
-            }
-
-            using (var subList = ImRaii.Child("TiltSubList", new Vector2(-1f, (float)(_activeSet.TiltReplacements.Count() * 30) + 30), false))
-            {
-                if (subList)
-                {
-                    foreach (var key in _activeSet.TiltReplacements.Keys)
-                    {
-                        var replacement = _activeSet.TiltReplacements[key].Replacement;
-
-                        if (ImGui.Checkbox("##" + key, ref _activeSet.TiltReplacements[key].Enabled))
-                        {
-                            Methods.SetupTilt(key);
-                            Service.Config.Save();
-                        }
-
-                        ImGui.SameLine();
-                        if (ImGui.Button(" - ##" + key))
-                        {
-                            _activeSet.TiltReplacements.Remove(key);
-                        }
-
-                        ImGui.SameLine();
-                        ImGui.Text($"#{key:D5}");
-
-                        ImGui.SameLine();
-                        if (_TiltitemWidth != 0 && Service.Config.AdvancedMode)
-                        {
-                            var widthRest = ImGui.GetWindowWidth() - _TiltitemWidth - ImGui.GetCursorPosX() - 5 * Scale;
-                            ImGui.PushTextWrapPos(Math.Max(widthRest, 60 * Scale) + ImGui.GetCursorPosX());
-                        }
-                        ImGui.TextWrapped(TiltReplacementsManager.GetName(key));
-                        if (_TiltitemWidth != 0 && Service.Config.AdvancedMode)
-                        {
-                            ImGui.PopTextWrapPos();
-                        }
-
-                        if (!Service.Config.AdvancedMode)
-                        {
-                            continue;
-                        }
-
-                        ImGui.SameLine();
-                        if (_TiltitemWidth != 0)
-                        {
-                            ImGui.SetCursorPosX(ImGui.GetWindowWidth() - _TiltitemWidth);
-                        }
-
-                        var startwidth = ImGui.GetCursorPosX();
-                        DrawItem("TiltRate", ref replacement.Unknown0, i => i.Unknown0);
-                        ImGui.SameLine();
-
-                        DrawItemByte("RotationOriginOffset", ref replacement.Unknown1, i => i.Unknown1);
-
-                        ImGui.SameLine();
-                        DrawItemByte("MaxAngle", ref replacement.Unknown2, i => i.Unknown2);
-
-                        ImGui.SameLine();
-                        DrawItemByte("Unknown3", ref replacement.Unknown3, i => i.Unknown3);
-
-                        ImGui.SameLine();
-                        DrawItemByte("Unknown4", ref replacement.Unknown4, i => i.Unknown4);
-
-                        ImGui.SameLine();
-                        DrawItemBool("MouseReverse", ref replacement.Unknown5, i => i.Unknown5);
-
-                        ImGui.SameLine();
-                        _TiltitemWidth = ImGui.GetCursorPosX() - startwidth;
-                        ImGui.NewLine();
-                        continue;
-
-                        void DrawItem(string name, ref ushort value,
-                            Func<Configurations.TiltReplacement, ushort> getDefault)
-                        {
-                            ImGui.SetNextItemWidth(60 * Scale);
-                            int relay = value;
-                            if (ImGui.DragInt("##" + name + key, ref relay))
-                            {
-                                value = (ushort)relay;
-                                Methods.SetupTilt(key);
-                                Service.Config.Save();
-                            }
-
-                            ImGui.SameLine();
-                            using (ImRaii.PushFont(UiBuilder.IconFont))
-                            {
-                                if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
-                                {
-                                    value = getDefault(TiltReplacementsManager.GetOriginalReplacement(key));
-                                    Methods.SetupTilt(key);
-                                    Service.Config.Save();
-                                }
-                            }
-                        }
-                        void DrawItemByte(string name, ref byte value,
-                            Func<Configurations.TiltReplacement, byte> getDefault)
-                        {
-                            ImGui.SetNextItemWidth(60 * Scale);
-                            int relay = value;
-                            if (ImGui.DragInt("##" + name + key, ref relay))
-                            {
-                                value = (byte)relay;
-                                Methods.SetupTilt(key);
-                                Service.Config.Save();
-                            }
-
-                            ImGui.SameLine();
-                            using (ImRaii.PushFont(UiBuilder.IconFont))
-                            {
-                                if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
-                                {
-                                    value = getDefault(TiltReplacementsManager.GetOriginalReplacement(key));
-                                    Methods.SetupTilt(key);
-                                    Service.Config.Save();
-                                }
-                            }
-                        }
-                        void DrawItemBool(string name, ref bool value,
-                            Func<Configurations.TiltReplacement, bool> getDefault)
-                        {
-                            ImGui.SetNextItemWidth(60 * Scale);
-                            bool relayBool = value;
-                            if (ImGui.Checkbox("##" + name + key, ref relayBool))
-                            {
-                                value = relayBool;
-                                Methods.SetupTilt(key);
-                                Service.Config.Save();
-                            }
-
-                            ImGui.SameLine();
-                            using (ImRaii.PushFont(UiBuilder.IconFont))
-                            {
-                                if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
-                                {
-                                    value = getDefault(TiltReplacementsManager.GetOriginalReplacement(key));
-                                    Methods.SetupTilt(key);
-                                    Service.Config.Save();
-                                }
-                            }
-                        }
-                    }
-
-                    const string searchTiltsPopup = "Search tilts";
-                    if (ImGui.Button(" + "))
-                    {
-                        ImGui.OpenPopup(searchTiltsPopup);
-                    }
-
-                    using var searchPopup = ImRaii.Popup(searchTiltsPopup);
-                    if (searchPopup)
-                    {
-                        var width = 200 * Scale;
-
-                        ImGui.SetNextItemWidth(width);
-                        ImGui.InputText("##Search Tilt", ref _searchTilt, 256);
-
-                        using var popUpChild = ImRaii.Child(searchTiltsPopup, new Vector2(width, 200 * Scale), true);
-                        foreach (var pair in TiltReplacementsManager.TiltNames.OrderBy(i =>
-                        {
-                            if (string.IsNullOrEmpty(_searchTilt)) return 0;
-                            return Math.Min(ScoreString(i.Value, _searchTilt),
-                                ScoreString(i.Key.ToString(), _searchTilt));
-                        }))
-                        {
-                            if (ImGui.Selectable($"#{pair.Key:D5} {pair.Value}"))
-                            {
-                                var original = TiltReplacementsManager.GetOriginalReplacement(pair.Key);
-                                _activeSet.TiltReplacements[pair.Key] =
-                                    new TiltReplacementConfig(new Configurations.TiltReplacement(
-                                            original.Unknown0,
-                                            original.Unknown1,
-                                            original.Unknown2,
-                                            original.Unknown3,
-                                            original.Unknown4,
-                                            original.Unknown5),
-                                        false);
-                                Service.Config.Save();
-                            }
-                        }
-                    }
-                }
-            }
-        }
     }
 
-    private string _searchAction = string.Empty;
-    private string _searchMount = string.Empty;
-    private string _searchTilt = string.Empty;
-
-    private static int ScoreString(string s1, string search)
+    public static int ScoreString(string s1, string search)
     {
         if (s1.Contains(search, StringComparison.OrdinalIgnoreCase))
         {
@@ -863,7 +307,7 @@ public sealed class ConfigWindow : Window
         return LevenshteinDistance(s1, search) + 20;
     }
 
-    private static int LevenshteinDistance(string s1, string s2)
+    public static int LevenshteinDistance(string s1, string s2)
     {
         var len1 = s1.Length;
         var len2 = s2.Length;
@@ -893,6 +337,7 @@ public sealed class ConfigWindow : Window
     {
         if (ImGui.Checkbox("Enable", ref Service.Config.EnableReplacement))
         {
+            //SETUP ALL PROCESSING
             Methods.SetupActions(ActionReplacementsManager.AllActionIds);
             Methods.SetupMounts(MountReplacementsManager.AllMountIds);
             Methods.SetupTilts(TiltReplacementsManager.AllTiltIds);
@@ -903,6 +348,7 @@ public sealed class ConfigWindow : Window
 
         if (ImGui.Button("Redraw"))
         {
+            //SETUP ALL PROCESSING
             Methods.SetupActions(ActionReplacementsManager.AllActionIds);
             Methods.SetupMounts(MountReplacementsManager.AllMountIds);
             Methods.SetupTilts(TiltReplacementsManager.AllTiltIds);
